requirements.txt
pandas>=2.0.0
pyyaml>=6.0.1
torch>=2.0.0,<=2.5.0